Add sample classes to separate layers
