+++ app/build.gradle
@@ -19,8 +19,8 @@ android {
         }
     }
     sourceSets {
-        main.kotlin.srcDirs += 'src/main/kotlin'
-        main.java.srcDirs += 'src/main/java'
+        main.java.srcDirs += 'src/main/kotlin'
+        test.java.srcDirs += 'src/test/kotlin'
     }
 }
