+++ app/src/main/kotlin/com/walterngo/wnproject/data/sampleapi/SampleApiClient.kt
@@ -1,0 +1,10 @@
+package com.walterngo.wnproject.data.sampleapi
+
+import com.walterngo.wnproject.domain.model.Sample
+
+class SampleApiClient {
+
+    fun getSample(): Sample {
+        return Sample("Hello World")
+    }
+}
