+++ app/src/main/kotlin/com/walterngo/wnproject/domain/model/Sample.kt
@@ -1,0 +1,3 @@
+package com.walterngo.wnproject.domain.model
+
+data class Sample(val message: String)
