+++ app/src/main/kotlin/com/walterngo/wnproject/domain/repository/SampleRepository.kt
@@ -1,0 +1,7 @@
+package com.walterngo.wnproject.domain.repository
+
+import com.walterngo.wnproject.domain.model.Sample
+
+interface SampleRepository {
+    fun getSample(): Sample
+}
