+++ app/src/main/kotlin/com/walterngo/wnproject/ui/main/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.walterngo.wnproject
+package com.walterngo.wnproject.ui.main
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import com.walterngo.wnproject.R
 
 class MainActivity : AppCompatActivity() {
