+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.walterngo.wnproject.MainActivity">
+    tools:context="com.walterngo.wnproject.main.kotlin.com.walterngo.wnproject.ui.main.MainActivity">
 
     <TextView
         android:layout_width="wrap_content"
